requirements.txt
pymongo==2.6.2
awesome-slugify==1.6.2
requests==2.6.0